do a more general implementation
the previous one relied on "the object with accession_x and hash_y" and
"the object with accession_x and version_z" as the same object. There
could be an object with accession_x, hash_y, version_w, which is not the
one that is going to be deleted, and a dupkey exception would be thrown.